pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-gate"
version = "0.3.0"
description = "LLM-Gate — AI-шлюз для обработки входящих инженерных задач"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.129.0",
    "uvicorn[standard]>=0.41.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.13.0",
    "jinja2>=3.1.6",
    "openai>=2.21.0",
    "tiktoken>=0.12.0",
    "faiss-cpu>=1.13.0",
    "sentence-transformers>=5.2.0",
    "mcp>=1.26.0",
    "qdrant-client>=1.16.0",
    "psycopg[binary,pool]>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
]

[project.scripts]

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["src/tests"]
pythonpath = ["src"]
markers = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]
filterwarnings = [
    "ignore:builtin type SwigPyPacked has no __module__ attribute:DeprecationWarning",
    "ignore:builtin type SwigPyObject has no __module__ attribute:DeprecationWarning",
    "ignore:builtin type swigvarlink has no __module__ attribute:DeprecationWarning",
]
